+++ pytorch_lightning/callbacks/__init__.py
@@ -27,7 +27,6 @@ from pytorch_lightning.callbacks.quantization import QuantizationAwareTraining
 from pytorch_lightning.callbacks.rich_model_summary import RichModelSummary
 from pytorch_lightning.callbacks.stochastic_weight_avg import StochasticWeightAveraging
 from pytorch_lightning.callbacks.timer import Timer
-from pytorch_lightning.callbacks.xla_stats_monitor import XLAStatsMonitor
 
 __all__ = [
     "BackboneFinetuning",
@@ -35,7 +34,6 @@ __all__ = [
     "Callback",
     "DeviceStatsMonitor",
     "EarlyStopping",
-    "XLAStatsMonitor",
     "GradientAccumulationScheduler",
     "LambdaCallback",
     "LearningRateMonitor",